chatbot-backend/requirements.txt
# Python Chatbot Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
sqlalchemy==2.0.25
pymysql==1.1.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.1.2
pillow==10.2.0
pypdf2==3.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.26.0
